Changed unsigned long to size_t as a recent Windows compiler warns about a conversion to long from a wider type. One applauds the compiler writer's advances in making code safer, but it does make for a bit of source adjustment from time to time. 	modified:   src/bin/dwarfdump/dd_safe_strcpy.c 	modified:   src/bin/dwarfdump/dd_safe_strcpy.h 	modified:   src/lib/libdwarf/dwarf_safe_strcpy.c 	modified:   src/lib/libdwarf/dwarf_safe_strcpy.h

+++ src/lib/libdwarf/dwarf_safe_strcpy.h
@@ -38,8 +38,8 @@ extern "C" {
 
 /*  The pointer arguments are required to be non-null.
     If outlen is zero nothing will be written to out. */
-void _dwarf_safe_strcpy(char *out, unsigned long outlen,
-    const char *in, unsigned long inlen);
+void _dwarf_safe_strcpy(char *out, size_t outlen,
+    const char *in, size_t inlen);
 
 #ifdef __cplusplus
 }
